Consolidating filter logic for integration tests
